Collapse loom test structure
