--- tests/loom/mod.rs
@@ -1,4 +1,0 @@
-mod vec;
-
-#[cfg(debug_assertions)]
-compile_error!("Loom tests are typically slow in debug mode. Run them with `--release`");
