--- tests/stub.rs
@@ -1,6 +1,0 @@
-#[cfg(loom)]
-/// Run loom tests with something like
-/// ```
-/// $ RUSTFLAGS="--cfg loom" cargo test --release loom
-/// ```
-mod loom;
